pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo Web Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "httpx>=0.25.2",
]
